Games fixes and updates

+++ play.it-2/games/play-the-inner-world.sh
@@ -1,8 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2019, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2020, Solène "Mopi" Huault
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,41 +31,56 @@ set -o errexit
 
 ###
 # The Inner World
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200206.3
 
 # Set game-specific variables
 
 GAME_ID='the-inner-world'
 GAME_NAME='The Inner World'
 
-ARCHIVES_LIST='ARCHIVE_GOG'
-
-ARCHIVE_GOG='setup_the_inner_world_2.0.0.2.exe'
+ARCHIVE_GOG='setup_the_inner_world_1.2.1_(26675).exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/the_inner_world'
-ARCHIVE_GOG_MD5='b5778aa9770ba7fc7d1a3884154c136b'
+ARCHIVE_GOG_MD5='2379e9c64c4fe06d6892474637aa405b'
 ARCHIVE_GOG_SIZE='1500000'
-ARCHIVE_GOG_VERSION='1.0-gog2.0.0.2'
+ARCHIVE_GOG_VERSION='1.2.1-gog26675'
 
-ARCHIVE_DOC_DATA_PATH='app'
-ARCHIVE_DOC_DATA_FILES='./eula?headup?games.rtf'
+ARCHIVE_GOG_OLD0='setup_the_inner_world_2.0.0.2.exe'
+ARCHIVE_GOG_OLD0_MD5='b5778aa9770ba7fc7d1a3884154c136b'
+ARCHIVE_GOG_OLD0_SIZE='1500000'
+ARCHIVE_GOG_OLD0_VERSION='1.2.1-gog2.0.0.2'
 
-ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./theinnerworld.exe ./adobe?air ./theinnerworld.swf'
+ARCHIVE_DOC_DATA_PATH='.'
+ARCHIVE_DOC_DATA_FILES='*.rtf'
+# Keep compatibility with old archives
+ARCHIVE_DOC_DATA_PATH_GOG_OLD0='app'
 
-ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./fizbin-link.url ./headup-link.url ./innerworld-link.url ./media ./meta-inf ./mimetype'
+ARCHIVE_GAME_BIN_PATH='.'
+ARCHIVE_GAME_BIN_FILES='*.exe *.swf adobe?air meta-inf mimetype'
+# Keep compatibility with old archives
+ARCHIVE_GAME_BIN_PATH_GOG_OLD0='app'
+
+ARCHIVE_GAME_DATA_PATH='.'
+ARCHIVE_GAME_DATA_FILES='*.url media'
+# Keep compatibility with old archives
+ARCHIVE_GAME_DATA_PATH_GOG_OLD0='app'
 
 APP_MAIN_TYPE='wine'
+# shellcheck disable=SC1004
 # shellcheck disable=SC2016
-APP_MAIN_PRERUN='[ "$(wine --version | sed "s/.\+(\(.\+\))/\\1/")" = "Staging" ] && winetricks csmt=on'
+APP_MAIN_PRERUN='# Enable CSMT patches on WINE staging
+if
+	[ "$(wine --version | sed "s/.\+(\(.\+\))/\\1/")" = "Staging" ] && \
+	[ ! -e csmt_enabled ]
+then
+	winetricks csmt=on
+	touch csmt_enabled
+fi'
 APP_MAIN_EXE='theinnerworld.exe'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='theinnerworld.exe'
-APP_MAIN_ICON_RES='16 32 48 128'
 
 PACKAGES_LIST='PKG_BIN PKG_DATA'
 
@@ -72,45 +88,52 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS="$PKG_DATA_ID wine"
+PKG_BIN_DEPS="$PKG_DATA_ID wine xcursor"
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"  "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
+# Get game icon
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
